# Avoid double cast in types that only override cast