--- a/activemodel/lib/active_model/type/boolean.rb
+++ b/activemodel/lib/active_model/type/boolean.rb
@@ -10,8 +10,6 @@ module ActiveModel
     # - Empty strings are coerced to +nil+.
     # - All other values will be coerced to +true+.
     class Boolean < Value
-      include SerializeCastValue
-
       FALSE_VALUES = [
         false, 0,
         "0", :"0",
@@ -29,6 +27,10 @@ module ActiveModel
 
       def serialize(value) # :nodoc:
         cast(value)
+      end
+
+      def serialize_cast_value(value) # :nodoc:
+        value
       end
 
       private