--- a/activemodel/lib/active_model/type/date.rb
+++ b/activemodel/lib/active_model/type/date.rb
@@ -22,7 +22,6 @@ module ActiveModel
     # String values are parsed using the ISO 8601 date format. Any other values
     # are cast using their +to_date+ method, if it exists.
     class Date < Value
-      include SerializeCastValue
       include Helpers::Timezone
       include Helpers::AcceptsMultiparameterTime.new