--- a/activemodel/lib/active_model/type/decimal.rb
+++ b/activemodel/lib/active_model/type/decimal.rb
@@ -32,7 +32,6 @@ module ActiveModel
     #     attribute :weight, :decimal, precision: 24
     #   end
     class Decimal < Value
-      include SerializeCastValue
       include Helpers::Numeric
       BIGDECIMAL_PRECISION = 18