--- a/activemodel/lib/active_model/type/string.rb
+++ b/activemodel/lib/active_model/type/string.rb
@@ -11,8 +11,6 @@ module ActiveModel
     # However, it accounts for mutable strings, so dirty tracking can properly
     # check if a string has changed.
     class String < ImmutableString
-      include SerializeCastValue
-
       def changed_in_place?(raw_old_value, new_value)
         if new_value.is_a?(::String)
           raw_old_value != new_value