--- a/activemodel/lib/active_model/type/time.rb
+++ b/activemodel/lib/active_model/type/time.rb
@@ -38,12 +38,11 @@ module ActiveModel
     #     attribute :start, :time, precision: 4
     #   end
     class Time < Value
-      include SerializeCastValue
       include Helpers::Timezone
-      include Helpers::TimeValue
       include Helpers::AcceptsMultiparameterTime.new(
         defaults: { 1 => 2000, 2 => 1, 3 => 1, 4 => 0, 5 => 0 }
       )
+      include Helpers::TimeValue
 
       def type
         :time
@@ -67,8 +66,6 @@ module ActiveModel
 
         super(value)
       end
-
-      alias :serialize_cast_value :serialize_time_value # :nodoc:
 
       private
         def cast_value(value)