--- a/activerecord/lib/active_record/type/date.rb
+++ b/activerecord/lib/active_record/type/date.rb
@@ -3,7 +3,6 @@
 module ActiveRecord
   module Type
     class Date < ActiveModel::Type::Date
-      include ActiveModel::Type::SerializeCastValue
       include Internal::Timezone
     end
   end