--- a/activerecord/lib/active_record/type/date_time.rb
+++ b/activerecord/lib/active_record/type/date_time.rb
@@ -3,7 +3,6 @@
 module ActiveRecord
   module Type
     class DateTime < ActiveModel::Type::DateTime
-      include ActiveModel::Type::SerializeCastValue
       include Internal::Timezone
     end
   end